+++ test/rtf_test.dart
@@ -7,17 +7,23 @@ import 'package:rtf/rtf.dart' as rtf;
 void main() {
   test('create a minimal document', () async {
     var el = [
-      rtf.Text('Test text',
-          style: rtf.TextStyle(style: 'heading 1', align: rtf.Align.right)),
+      rtf.Paragraph(style: 'Heading 1', align: rtf.Align.right, children: [
+        rtf.Text('Test text',
+            style: rtf.TextStyle(variations: [rtf.StyleVariation.italic])),
+        rtf.Text(' Continuos test text',
+            style: rtf.TextStyle(
+                font: 'Normal', variations: [rtf.StyleVariation.underline]))
+      ]),
       rtf.NewLine(),
-      rtf.Text('Second test text', style: rtf.TextStyle(style: 'Normal')),
+      rtf.Paragraph(style: 'Normal', children: [rtf.Text('Second test text')]),
     ];
-    rtf.Document doc = rtf.Document(el, hdLeft: rtf.PageNo());
-    doc.addFont('Normal', 'roman', 'Times New Roman', rtf.FontStyle.regular, 9);
-    doc.addFont(
-        'heading 1', 'roman', 'Times New Roman', rtf.FontStyle.bold, 14);
-    doc.addFont(
-        'heading 2', 'roman', 'Times New Roman', rtf.FontStyle.bold, 12);
+    rtf.Document doc = rtf.Document(el, hdLeft: rtf.PageNo(), styles: [
+      rtf.Style('Normal', rtf.FontFamily.roman, 'Times New Roman', 9),
+      rtf.Style('Heading 1', rtf.FontFamily.roman, 'Times New Roman', 14,
+          [rtf.StyleVariation.bold]),
+      rtf.Style('Heading 2', rtf.FontFamily.roman, 'Times New Roman', 12,
+          [rtf.StyleVariation.bold])
+    ]);
     await doc.save(File('../test.rtf'));
   });
   test('draw a Line', () async {
@@ -28,12 +34,13 @@ void main() {
       rtf.NewLine(),
       rtf.Listing([rtf.Text('a'), rtf.Text('b')], true),
       rtf.Listing([rtf.Text('A'), rtf.Text('B')], false)
+    ], styles: [
+      rtf.Style('Normal', rtf.FontFamily.roman, 'Times New Roman', 9),
+      rtf.Style('Heading 1', rtf.FontFamily.roman, 'Times New Roman', 14,
+          [rtf.StyleVariation.bold]),
+      rtf.Style('Heading 2', rtf.FontFamily.roman, 'Times New Roman', 12,
+          [rtf.StyleVariation.bold])
     ]);
-    doc.addFont('Normal', 'roman', 'Times New Roman', rtf.FontStyle.regular, 9);
-    doc.addFont(
-        'heading 1', 'roman', 'Times New Roman', rtf.FontStyle.bold, 14);
-    doc.addFont(
-        'heading 2', 'roman', 'Times New Roman', rtf.FontStyle.bold, 12);
     await doc.save(File('../line.rtf'));
   });
   test('draw a Row', () async {
@@ -46,12 +53,13 @@ void main() {
             children: List.generate(100, (index) => rtf.Text('jjjj'))),
         rtf.Text('fff')
       ]),
+    ], styles: [
+      rtf.Style('Normal', rtf.FontFamily.roman, 'Times New Roman', 9),
+      rtf.Style('Heading 1', rtf.FontFamily.roman, 'Times New Roman', 14,
+          [rtf.StyleVariation.bold]),
+      rtf.Style('Heading 2', rtf.FontFamily.roman, 'Times New Roman', 12,
+          [rtf.StyleVariation.bold])
     ]);
-    doc.addFont('Normal', 'roman', 'Times New Roman', rtf.FontStyle.regular, 9);
-    doc.addFont(
-        'heading 1', 'roman', 'Times New Roman', rtf.FontStyle.bold, 14);
-    doc.addFont(
-        'heading 2', 'roman', 'Times New Roman', rtf.FontStyle.bold, 12);
     await doc.save(File('../row.rtf'));
   });
   test('write a table', () async {
@@ -60,29 +68,36 @@ void main() {
         (index) => List.generate(
             4,
             (c) => rtf.Text('Cell$index$c',
-                style: rtf.TextStyle(style: 'Normal'))));
+                style: rtf.TextStyle(font: 'Normal'))));
     c.first.first = rtf.Column(lastNL: false, children: [
-      rtf.Text('1', style: rtf.TextStyle(style: 'Normal')),
-      rtf.Text('alfa', style: rtf.TextStyle(style: 'Normal'))
+      rtf.Text('1', style: rtf.TextStyle(font: 'Normal')),
+      rtf.Text('alfa', style: rtf.TextStyle(font: 'Normal'))
     ]);
     c.last.removeLast();
     c.last[1] = rtf.ColSpan(2,
-        child: rtf.Text('Span', style: rtf.TextStyle(align: rtf.Align.center)));
+        child: rtf.Paragraph(
+            style: 'Normal',
+            align: rtf.Align.center,
+            children: [rtf.Text('Span')]));
     var el = [
       rtf.Table(
           [
-            rtf.Text('First column',
-                style:
-                    rtf.TextStyle(style: 'heading 2', align: rtf.Align.center)),
-            rtf.Text('Second column',
-                style:
-                    rtf.TextStyle(style: 'heading 2', align: rtf.Align.center)),
-            rtf.Text('Third column',
-                style:
-                    rtf.TextStyle(style: 'heading 2', align: rtf.Align.center)),
-            rtf.Text('Fourth column',
-                style:
-                    rtf.TextStyle(style: 'heading 2', align: rtf.Align.center)),
+            rtf.Paragraph(
+                style: 'Heading 2',
+                align: rtf.Align.center,
+                children: [rtf.Text('First column')]),
+            rtf.Paragraph(
+                style: 'Heading 2',
+                align: rtf.Align.center,
+                children: [rtf.Text('Second column')]),
+            rtf.Paragraph(
+                style: 'Heading 2',
+                align: rtf.Align.center,
+                children: [rtf.Text('Third column')]),
+            rtf.Paragraph(
+                style: 'Heading 2',
+                align: rtf.Align.center,
+                children: [rtf.Text('Fourth column')]),
           ],
           c,
           colWidths: [100, 150, 100, 100],
@@ -95,10 +110,13 @@ void main() {
           top: rtf.TableBorder.standard(),
           bottom: rtf.TableBorder.standard())
     ];
-    rtf.Document doc = rtf.Document(el);
-    doc.addFont('Normal', 'swiss', 'Arial', rtf.FontStyle.regular, 9);
-    doc.addFont('heading 1', 'swiss', 'Arial', rtf.FontStyle.bold, 14);
-    doc.addFont('heading 2', 'swiss', 'Arial', rtf.FontStyle.bold, 12);
+    rtf.Document doc = rtf.Document(el, styles: [
+      rtf.Style('Normal', rtf.FontFamily.swiss, 'Arial', 9),
+      rtf.Style('Heading 1', rtf.FontFamily.swiss, 'Arial', 14,
+          [rtf.StyleVariation.bold]),
+      rtf.Style('Heading 2', rtf.FontFamily.swiss, 'Arial', 12,
+          [rtf.StyleVariation.bold])
+    ]);
     await doc.save(File('../table.rtf'));
   });
 }